+++ libraries/base/changelog.md
@@ -62,6 +62,10 @@
     super-class of `Monoid` in the future). These modules were
     provided by the `semigroups` package previously. (#10365)
 
+  * Add `URec`, `UAddr`, `UChar`, `UDouble`, `UFloat`, `UInt`, and `UWord` to
+    `GHC.Generics` as part of making GHC generics capable of handling
+    unlifted types (#10868)
+
 ## 4.8.1.0  *Jul 2015*
 
   * Bundled with GHC 7.10.2